requirements.txt
streamlit
pdfminer.six
pandas
scikit-learn
joblib
tqdm
